refactor : method transfer
