--- src/main/java/RacingClient.java
@@ -1,7 +1,0 @@
-import view.InputView;
-import view.ResultView;
-
-public class RacingClient {
-    public static void main(String[] args) {
-    }
-}
