+++ src/main/java/model/Car.java
@@ -1,5 +1,7 @@
 package model;
 
+import java.util.ArrayList;
+
 public class Car{
     private String name;
     private int position = 0;
@@ -18,5 +20,16 @@ public class Car{
 
     public void increasePosition(){
         this.position++;
+    }
+
+    public static ArrayList<Car> generateCars(String[] carNames) throws Exception{
+        ArrayList<Car> cars = new ArrayList<>();
+        for (String carName : carNames){
+            if (carName.length() > 6) {
+                throw new Exception();
+            }
+            cars.add(new Car(carName));
+        }
+        return cars;
     }
 }
