+++ src/main/java/view/ResultView.java
@@ -1,8 +1,23 @@
 package view;
 
+import controller.RacingGame;
+import model.Car;
+
+import java.util.Collections;
+
 public class ResultView{
 
-    public void finalResult(){
+    private RacingGame racingGame = new RacingGame();
+
+    public void printTrialResult(){
+        for (Car car : racingGame.getCars()){
+            System.out.print(car.getName() + " : " + String.join("", Collections.nCopies(car.getPosition(), "*")) + "\n");
+        }
+        System.out.println();
+    }
+
+    public void printFinalWinner(){
+        System.out.print(String.join(",",  racingGame.searchWinners()));
         System.out.println("가 최종 우승했습니다.");
     }
 }
