+++ src/test/java/InputViewTest.java
@@ -1,0 +1,12 @@
+import static org.junit.Assert.*;
+
+public class InputViewTest {
+
+    @org.junit.Test
+    public void inputCarName() {
+    }
+
+    @org.junit.Test
+    public void inputNumberOfTrials() {
+    }
+}
